requirements.txt
Flask
prometheus_client
PyYAML
requests
urllib3
